 invest-robot-contest_tinkoff-contest-python-main/requirements.txt
pyyaml
tinkoff-investments
